Split up the java library into 3 modules
This splits the java library into three separate modules:
- adhan-core (the core library)
- adhan (adhan-core + threetenbp)
- adhan-android (adhan-core + android threetenabp)

We can't just use threetenbp on Android because of a bug that causes
Android to load the entire set of information into memory. The android
version of the library includes threetenbp, but modifies it so that this
doesn't happen.

Consequently, adhan-android and adhan are (almost empty) modules that
just include adhan-core plus the correct dependency.