Use java6 time apis instead of JSR-310 backport
The JSR-310 backport api is definitely a better api than the Java 6
calendar apis. However, due to excessive caching with resource loading
on Android (used for loading the timezone database), using this library
on Android should ideally use an Android wrapped version instead (one
that loads the timezone data from AssetManager).

Implementing this correctly in Adhan would have therefore required
splitting up this library into 3 modules - a core module, a library
module (for use by Java projects depending on threeten backport), and an
Android module (outputting an aar for use by Android projects
and  depending on the Android specific flavor of threeten backport).

Since this is a hassle and costs the developer ~3000 methods towards the
method limit, and since we don't really deal with timezone conversions
within the library (everything is done in UTC), the alternative was to
just use the java 6 apis. While more painful for the library developer,
this simplifies life for the application developers and allows them to
use whatever time library they want in their app (be it joda, threeten
backport, or even native java 8 when not on Android).